handle sending via a list of provided relays
these relays can either be connected or not,
we have to check, and optionally connect them via the
`Nostrbase.add_relay/1` async function, hence the async handling.

this is really only for cases when the user provides an nprofile or
nevent type entity that we have to search for without being connected
to first.