signer: change module load order, dispatcher waits, and simplify profile setting
the Signer depends on the Repo being started, so we move it to after in
the start sequence.
The Dispatcher waits for the Signer to be ready, before returning a
profile or not. This is the time the Signer's handle_continue clause
runs.
Move profile setting as persistent_term and db to the Signer module.